+++ app/src/main/res/layout/layout_title_item.xml
@@ -1,9 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:paddingHorizontal="16dp"
-    android:paddingTop="12dp">
+    android:layout_height="wrap_content">
 
     <TextView
         android:id="@+id/title"
